# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API for Timeful.",
        "title": "Timeful API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3002/api",
    "paths": {
        "/analytics/monthly-active-event-creators": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Gets the daily count of monthly active event creators over a date range",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start date (YYYY-MM-DD) for the range",
                        "name": "startDate",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End date (YYYY-MM-DD) for the range",
                        "name": "endDate",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Client's timezone offset in minutes from UTC (e.g., -420 for UTC-7)",
                        "name": "timezoneOffset",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "count": {
                                        "type": "integer"
                                    },
                                    "date": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid date format, range, or timezone offset",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/analytics/monthly-active-event-creators-with-more-than-x-events": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Gets the daily count of monthly active event creators over a date range with more than x events",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start date (YYYY-MM-DD) for the range",
                        "name": "startDate",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End date (YYYY-MM-DD) for the range",
                        "name": "endDate",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Client's timezone offset in minutes from UTC (e.g., -420 for UTC-7)",
                        "name": "timezoneOffset",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "count": {
                                        "type": "integer"
                                    },
                                    "date": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid date format, range, or timezone offset",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/analytics/scanned-poster": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Notifies us when poster QR code has been scanned",
                "parameters": [
                    {
                        "description": "Object containing the location that poster was scanned from and the url that was scanned",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "location": {
                                    "$ref": "#/definitions/models.Location"
                                },
                                "url": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/analytics/user/{email}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Gets the user by email",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User email",
                        "name": "email",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.AppleCalendarAuth": {
            "type": "object"
        },
        "models.BufferTimeOptions": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean"
                },
                "time": {
                    "type": "integer"
                }
            }
        },
        "models.CalendarAccount": {
            "type": "object",
            "properties": {
                "appleCalendarAuth": {
                    "$ref": "#/definitions/models.AppleCalendarAuth"
                },
                "calendarType": {
                    "$ref": "#/definitions/models.CalendarType"
                },
                "email": {
                    "description": "Email is required for all calendar accounts",
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "icsCalendarAuth": {
                    "$ref": "#/definitions/models.ICSCalendarAuth"
                },
                "oAuth2CalendarAuth": {
                    "$ref": "#/definitions/models.OAuth2CalendarAuth"
                },
                "picture": {
                    "type": "string"
                },
                "subCalendars": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/models.SubCalendar"
                    }
                }
            }
        },
...
        "models.CalendarOptions": {
            "type": "object",
            "properties": {
                "bufferTime": {
                    "$ref": "#/definitions/models.BufferTimeOptions"
                },
                "workingHours": {
                    "$ref": "#/definitions/models.WorkingHoursOptions"
                }
            }
        },
        "models.CalendarType": {
            "type": "string",
            "enum": [
                "apple",
                "google",
                "outlook",
                "ics"
            ],
            "x-enum-varnames": [
                "AppleCalendarType",
                "GoogleCalendarType",
                "OutlookCalendarType",
                "ICSCalendarType"
            ]
        },
...
        "models.ICSCalendarAuth": {
            "type": "object",
            "properties": {
                "label": {
                    "type": "string"
                }
            }
        },
        "models.Location": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "country_code": {
                    "type": "string"
                },
                "country_name": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "postal": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                }
            }
        },
        "models.OAuth2CalendarAuth": {
            "type": "object"
        },
...
        "models.SubCalendar": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "_id": {
                    "description": "Profile info",
                    "type": "string"
                },
                "calendarAccounts": {
                    "description": "CalendarAccounts is a mapping from {`email_CALENDARTYPE` =\u003e CalendarAccount} that contains all the\nadditional accounts the user wants to see google calendar events for",
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/models.CalendarAccount"
                    }
                },
                "calendarOptions": {
                    "description": "Calendar options",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.CalendarOptions"
                        }
                    ]
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "hasCustomName": {
                    "description": "Whether the user has set a custom name for themselves, i.e. don't change their name when they sign in",
                    "type": "boolean"
                },
                "lastName": {
                    "type": "string"
                },
                "numEventsCreated": {
                    "type": "integer"
                },
                "picture": {
                    "type": "string"
                },
                "primaryAccountKey": {
                    "description": "The calendarAccountKey of the account the user first signed in with",
                    "type": "string"
                },
                "timezoneOffset": {
                    "type": "integer"
                }
            }
        },
        "models.WorkingHoursOptions": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean"
                },
                "endTime": {
                    "type": "number"
                },
                "startTime": {
                    "type": "number"
                }
            }
        },
...
    }
}